fix: fix problem with the connection of the data base and the data management of entires

+++ src/DistributionCenter.Domain/Entities/Concretes/Product.cs
@@ -6,5 +6,5 @@ public class Product : BaseEntity
 {
     public required string Name { get; set; }
     public required string Description { get; set; }
-    public required double Weight { get; set; }
+    public required int Weight { get; set; }
 }
